optimize FileDbUtils; optimize get folder path by date method

+++ xuecheng-edu-media/xuecheng-edu-media-service/src/main/java/com/xuecheng/media/utils/FileDbUtils.java
@@ -42,6 +42,11 @@ public class FileDbUtils {
      */
     @Transactional(rollbackFor = Exception.class)
     public MediaFile addFileInfo(Long companyId, String fileMd5, FileParamsDto dto, String bucket, String filename, String objectName) {
+        if (fileMd5 == null || fileMd5.trim().isEmpty()) {
+            log.error("传入的文件 MD5 值为空, objectName={}, FileParamsDto={}", objectName, dto);
+            throw new RuntimeException("传入的文件 MD5 值为空");
+        }
+
         // 查询数据库中是否已存在该文件
         MediaFile mediaFile = getOneFileInfo(fileMd5);
 
@@ -65,6 +70,11 @@ public class FileDbUtils {
      */
     @Transactional(rollbackFor = Exception.class)
     public MediaFile updateFileInfo(Long companyId, String fileMd5, FileParamsDto dto, String bucket, String filename, String objectName) {
+        if (fileMd5 == null || fileMd5.trim().isEmpty()) {
+            log.error("传入的文件 MD5 值为空, objectName={}, FileParamsDto={}", objectName, dto);
+            throw new RuntimeException("传入的文件 MD5 值为空");
+        }
+
         // 查询数据库中是否存在该文件
         MediaFile mediaFile = getOneFileInfo(fileMd5);
 
@@ -73,9 +83,9 @@ public class FileDbUtils {
             return saveFileInfo(companyId, fileMd5, dto, bucket, filename, objectName);
         } else {
             log.error("尝试更新数据库中不存在的文件信息, objectName={}, fileMd5={}", objectName, fileMd5);
+            XueChengEduException.cast("尝试更新数据库中不存在的文件信息");
+            return null;
         }
-
-        return null;
     }
 
     /**
@@ -84,34 +94,39 @@ public class FileDbUtils {
      */
     @Transactional(rollbackFor = Exception.class)
     public boolean deleteFileInfo(MediaFile file) {
-        int res = mediaFileMapper.delete(new LambdaQueryWrapper<MediaFile>()
-                // 机构 ID
-                .eq(file.getCompanyId() != null, MediaFile::getCompanyId, file.getCompanyId())
-                // 文件名
-                .eq(StringUtils.isNotEmpty(file.getFilename()), MediaFile::getFilename, file.getFilename())
-                // 文件类型
-                .eq(StringUtils.isNotEmpty(file.getFileType()), MediaFile::getFileType, file.getFileType())
-                // 文件存储路径
-                .eq(StringUtils.isNotEmpty(file.getFilePath()), MediaFile::getFilePath, file.getFilePath())
-                // 文件所在桶名
-                .eq(StringUtils.isNotEmpty(file.getBucket()), MediaFile::getBucket, file.getBucket())
-                // 文件大小
-                .eq(file.getFileSize() != null, MediaFile::getFileSize, file.getFileSize())
-                // 文件标签
-                .eq(StringUtils.isNotEmpty(file.getTags()), MediaFile::getTags, file.getTags())
-                // 文件上传人名称
-                .eq(StringUtils.isNotEmpty(file.getUsername()), MediaFile::getUsername, file.getUsername())
-                // 文件备注
-                .eq(StringUtils.isNotEmpty(file.getRemark()), MediaFile::getRemark, file.getRemark()));
+        if (file != null) {
+            int res = mediaFileMapper.delete(new LambdaQueryWrapper<MediaFile>()
+                    // 机构 ID
+                    .eq(file.getCompanyId() != null, MediaFile::getCompanyId, file.getCompanyId())
+                    // 文件名
+                    .eq(StringUtils.isNotEmpty(file.getFilename()), MediaFile::getFilename, file.getFilename())
+                    // 文件类型
+                    .eq(StringUtils.isNotEmpty(file.getFileType()), MediaFile::getFileType, file.getFileType())
+                    // 文件存储路径
+                    .eq(StringUtils.isNotEmpty(file.getFilePath()), MediaFile::getFilePath, file.getFilePath())
+                    // 文件所在桶名
+                    .eq(StringUtils.isNotEmpty(file.getBucket()), MediaFile::getBucket, file.getBucket())
+                    // 文件大小
+                    .eq(file.getFileSize() != null, MediaFile::getFileSize, file.getFileSize())
+                    // 文件标签
+                    .eq(StringUtils.isNotEmpty(file.getTags()), MediaFile::getTags, file.getTags())
+                    // 文件上传人名称
+                    .eq(StringUtils.isNotEmpty(file.getUsername()), MediaFile::getUsername, file.getUsername())
+                    // 文件备注
+                    .eq(StringUtils.isNotEmpty(file.getRemark()), MediaFile::getRemark, file.getRemark()));
 
-        if (res > 0) {
-            log.debug("媒资文件删除成功, objectName={}", file.getFilePath());
-            return true;
+            if (res > 0) {
+                log.debug("媒资文件删除成功, objectName={}", file.getFilePath());
+                return true;
+            } else {
+                log.error("从数据库删除文件信息失败, objectName={}", file.getFilePath());
+                XueChengEduException.cast("从数据库删除文件信息失败，未知错误!");
+            }
         } else {
-            log.error("从数据库删除文件信息失败, objectName={}", file.getFilePath());
-            XueChengEduException.cast("从数据库删除文件信息失败，未知错误!");
-            return false;
+            log.error("删除文件信息失败, 传入的文件信息为空");
+            throw new RuntimeException("传入的文件信息为空");
         }
+        return false;
     }
 
     /**
@@ -123,25 +138,30 @@ public class FileDbUtils {
      * @return 媒体/视频 文件信息的 {@link List}
      */
     public List<MediaFile> getListFileInfo(Long companyId, FileParamsDto dto, String bucket, String objectName) {
-        return mediaFileMapper.selectList(new LambdaQueryWrapper<MediaFile>()
-                // 机构 ID
-                .eq(companyId != null, MediaFile::getCompanyId, companyId)
-                // 文件名
-                .eq(StringUtils.isNotEmpty(dto.getFilename()), MediaFile::getFilename, dto.getFilename())
-                // 文件类型
-                .eq(StringUtils.isNotEmpty(dto.getFileType()), MediaFile::getFileType, dto.getFileType())
-                // 文件存储路径
-                .eq(objectName != null, MediaFile::getFilePath, objectName)
-                // 文件访问 url
-                .eq(objectName != null, MediaFile::getUrl, "/" + bucket + "/" + objectName)
-                // 文件大小
-                .eq(dto.getFileSize() != null, MediaFile::getFileSize, dto.getFileSize())
-                // 文件标签
-                .eq(StringUtils.isNotEmpty(dto.getTags()), MediaFile::getTags, dto.getTags())
-                // 文件上传人名称
-                .eq(StringUtils.isNotEmpty(dto.getUsername()), MediaFile::getUsername, dto.getUsername())
-                // 文件备注
-                .eq(StringUtils.isNotEmpty(dto.getRemark()), MediaFile::getRemark, dto.getRemark()));
+        if (dto != null) {
+            return mediaFileMapper.selectList(new LambdaQueryWrapper<MediaFile>()
+                    // 机构 ID
+                    .eq(companyId != null, MediaFile::getCompanyId, companyId)
+                    // 文件名
+                    .eq(StringUtils.isNotEmpty(dto.getFilename()), MediaFile::getFilename, dto.getFilename())
+                    // 文件类型
+                    .eq(StringUtils.isNotEmpty(dto.getFileType()), MediaFile::getFileType, dto.getFileType())
+                    // 文件存储路径
+                    .eq(objectName != null, MediaFile::getFilePath, objectName)
+                    // 文件访问 url
+                    .eq(objectName != null, MediaFile::getUrl, "/" + bucket + "/" + objectName)
+                    // 文件大小
+                    .eq(dto.getFileSize() != null, MediaFile::getFileSize, dto.getFileSize())
+                    // 文件标签
+                    .eq(StringUtils.isNotEmpty(dto.getTags()), MediaFile::getTags, dto.getTags())
+                    // 文件上传人名称
+                    .eq(StringUtils.isNotEmpty(dto.getUsername()), MediaFile::getUsername, dto.getUsername())
+                    // 文件备注
+                    .eq(StringUtils.isNotEmpty(dto.getRemark()), MediaFile::getRemark, dto.getRemark()));
+        } else {
+            log.error("查询文件列表失败, 传入的 FileParamsDto 为空");
+            throw new RuntimeException("传入的 FileParamsDto 为空");
+        }
     }
 
     /**
@@ -150,52 +170,62 @@ public class FileDbUtils {
      * @return 文件信息对象 {@link MediaFile}，若不存在则返回 null
      */
     public MediaFile getOneFileInfo(String fileMd5) {
+        if (fileMd5 == null || fileMd5.trim().isEmpty()) {
+            log.error("传入的文件 MD5 值为空");
+            throw new RuntimeException("传入的文件 MD5 值为空");
+        }
+
         return mediaFileMapper.selectOne(new LambdaQueryWrapper<MediaFile>().eq(MediaFile::getId, fileMd5));
     }
 
     private MediaFile saveFileInfo(Long companyId, String fileMd5, FileParamsDto dto, String bucket, String filename, String objectName) {
-        MediaFile mediaFile = new MediaFile();
-        BeanUtils.copyProperties(dto, mediaFile);
+        if (fileMd5 != null && !fileMd5.trim().isEmpty()) {
+            MediaFile mediaFile = new MediaFile();
+            BeanUtils.copyProperties(dto, mediaFile);
 
-        // 文件 ID
-        mediaFile.setId(fileMd5);
-        // 机构 ID
-        mediaFile.setCompanyId(companyId);
-        // 机构名称
-        // mediaFile.setCompanyName("");
-        // 文件名
-        mediaFile.setFilename(filename);
-        // 标签
-        // mediaFile.setTags("");
-        // 备注
-        // mediaFile.setRemark("");
-        // 桶
-        mediaFile.setBucket(bucket);
-        // 文件存储路径
-        mediaFile.setFilePath(objectName);
-        // 文件标识 ID
-        mediaFile.setFileId(fileMd5);
-        // 文件访问 url
-        mediaFile.setUrl("/" + bucket + "/" + objectName);
-        // 上传人
-        // mediaFile.setUsername("");
-        // 上传时间
-        mediaFile.setCreateDate(LocalDateTime.now());
-        // 状态
-        mediaFile.setStatus("1");
-        // 审核状态：审核通过
-        mediaFile.setAuditStatus("002003");
+            // 文件 ID
+            mediaFile.setId(fileMd5);
+            // 机构 ID
+            mediaFile.setCompanyId(companyId);
+            // 机构名称
+            // mediaFile.setCompanyName("");
+            // 文件名
+            mediaFile.setFilename(filename);
+            // 标签
+            // mediaFile.setTags("");
+            // 备注
+            // mediaFile.setRemark("");
+            // 桶
+            mediaFile.setBucket(bucket);
+            // 文件存储路径
+            mediaFile.setFilePath(objectName);
+            // 文件标识 ID
+            mediaFile.setFileId(fileMd5);
+            // 文件访问 url
+            mediaFile.setUrl("/" + bucket + "/" + objectName);
+            // 上传人
+            // mediaFile.setUsername("");
+            // 上传时间
+            mediaFile.setCreateDate(LocalDateTime.now());
+            // 状态
+            mediaFile.setStatus("1");
+            // 审核状态：审核通过
+            mediaFile.setAuditStatus("002003");
 
-        // 插入数据库
-        int res = mediaFileMapper.insert(mediaFile);
+            // 插入数据库
+            int res = mediaFileMapper.insert(mediaFile);
 
-        if (res <= 0) {
-            log.error("文件信息保存到数据表 media_file 失败, bucket={}, objectName={}", bucket, objectName);
-            XueChengEduException.cast("文件上传后保存信息失败");
-            return null;
+            if (res <= 0) {
+                log.error("文件信息保存到数据表 media_file 失败, bucket={}, objectName={}", bucket, objectName);
+                XueChengEduException.cast("文件上传后保存信息失败");
+                return null;
+            }
+            log.debug("文件信息保存到数据表 media_file 成功, fileMd5={}, objectName={}", fileMd5, objectName);
+            return mediaFile;
+        } else {
+            log.error("添加文件信息失败, 传入的文件 MD5 为空");
+            throw new RuntimeException("传入的文件 MD5 值为空");
         }
-        log.debug("文件信息保存到数据表 media_file 成功, fileMd5={}, objectName={}", fileMd5, objectName);
-        return mediaFile;
     }
 
 }
